format solidity files fix

+++ test/unit/FundMeTest.t.sol
@@ -2,9 +2,9 @@
 
 pragma solidity ^0.8.19;
 
-import {Test, console} from 'forge-std/Test.sol';
-import {FundMe} from '../../src/FundMe.sol';
-import {FundMeDeployScript} from '../../script/FundMeDeployScript.s.sol';
+import {Test, console} from "forge-std/Test.sol";
+import {FundMe} from "../../src/FundMe.sol";
+import {FundMeDeployScript} from "../../script/FundMeDeployScript.s.sol";
 
 contract FundMeTest is Test {
     FundMe fundMe;
@@ -15,77 +15,84 @@ contract FundMeTest is Test {
     uint256 GAS_PRICE = 1;
 
     function setUp() external {
-       // fundMe = new FundMe(0x694AA1769357215DE4FAC081bf1f309aDC325306);
-       FundMeDeployScript Deployer = new FundMeDeployScript();
-       fundMe = Deployer.run();
-       vm.deal(USER, STARTING_BALANCE);
+        // fundMe = new FundMe(0x694AA1769357215DE4FAC081bf1f309aDC325306);
+        FundMeDeployScript Deployer = new FundMeDeployScript();
+        fundMe = Deployer.run();
+        vm.deal(USER, STARTING_BALANCE);
     }
+
     function testingMinimumUsd() public view {
         assertEq(fundMe.MINIMUM_USD(), 5e18);
     }
+
     function testOwner() public view {
         assertEq(fundMe.getOwner(), msg.sender);
     }
+
     function testPriceFeedVersionAccuracy() public view {
-        if(block.chainid == 11155111) {
+        if (block.chainid == 11155111) {
             assertEq(fundMe.getVersion(), 4);
-        }
-        else if (block.chainid == 1) {
+        } else if (block.chainid == 1) {
             assertEq(fundMe.getVersion(), 6);
-        } 
+        }
     }
+
     function testMinimum_USDFailure() public {
         vm.expectRevert();
         fundMe.fund();
     }
+
     modifier funded() {
         vm.prank(USER);
         fundMe.fund{value: TEST_AMOUNT}();
         _;
     }
+
     function testFundStructure() public funded {
         uint256 amountFunded = fundMe.getAddressToAmountFunded(USER);
         assertEq(amountFunded, TEST_AMOUNT);
     }
+
     function testAddFunderToArray() public funded {
         address funder = fundMe.getFunder(0);
         assertEq(funder, USER);
     }
+
     function testOnlyOwnercanWithdraw() public funded {
         vm.expectRevert();
         vm.prank(USER);
         fundMe.withdraw();
     }
+
     function testWithdrawFromSingleFunder() public funded {
         //arrange
-        uint startingOwnerBalance = fundMe.getOwner().balance;
-        uint startingFundMeBalance = address(fundMe).balance;
+        uint256 startingOwnerBalance = fundMe.getOwner().balance;
+        uint256 startingFundMeBalance = address(fundMe).balance;
 
         //act
         vm.txGasPrice(GAS_PRICE);
         vm.prank(fundMe.getOwner());
         fundMe.withdraw();
 
-        
         //assert
-        uint endingOwnerBalance = fundMe.getOwner().balance;
-        uint endingFundMeBalance = address(fundMe).balance;
+        uint256 endingOwnerBalance = fundMe.getOwner().balance;
+        uint256 endingFundMeBalance = address(fundMe).balance;
 
         assertEq(endingFundMeBalance, 0);
         assertEq(startingFundMeBalance + startingOwnerBalance, endingOwnerBalance);
     }
+
     function testWithdrawFromMultipleFunders() public funded {
         uint160 numberOfFunders = 10;
         uint160 startingFunderIndex = 1;
 
-        for (uint160 i = startingFunderIndex; i < numberOfFunders; i++){
+        for (uint160 i = startingFunderIndex; i < numberOfFunders; i++) {
             hoax(address(i), TEST_AMOUNT);
             fundMe.fund{value: TEST_AMOUNT}();
         }
-        
 
-        uint startingOwnerBalance = fundMe.getOwner().balance;
-        uint startingFundMeBalance = address(fundMe).balance;
+        uint256 startingOwnerBalance = fundMe.getOwner().balance;
+        uint256 startingFundMeBalance = address(fundMe).balance;
 
         vm.startPrank(fundMe.getOwner());
         fundMe.withdraw();
@@ -94,18 +101,18 @@ contract FundMeTest is Test {
         assertEq(address(fundMe).balance, 0);
         assertEq(startingFundMeBalance + startingOwnerBalance, fundMe.getOwner().balance);
     }
+
     function testWithdrawFromMultipleFundersCheaper() public funded {
         uint160 numberOfFunders = 10;
         uint160 startingFunderIndex = 1;
 
-        for (uint160 i = startingFunderIndex; i < numberOfFunders; i++){
+        for (uint160 i = startingFunderIndex; i < numberOfFunders; i++) {
             hoax(address(i), TEST_AMOUNT);
             fundMe.fund{value: TEST_AMOUNT}();
         }
-        
 
-        uint startingOwnerBalance = fundMe.getOwner().balance;
-        uint startingFundMeBalance = address(fundMe).balance;
+        uint256 startingOwnerBalance = fundMe.getOwner().balance;
+        uint256 startingFundMeBalance = address(fundMe).balance;
 
         vm.startPrank(fundMe.getOwner());
         fundMe.cheaperWithdraw();
@@ -114,4 +121,4 @@ contract FundMeTest is Test {
         assertEq(address(fundMe).balance, 0);
         assertEq(startingFundMeBalance + startingOwnerBalance, fundMe.getOwner().balance);
     }
-} 
+}
